[Suraj]: Update the auth-service functionality

+++ auth-service/src/main/java/com/stockpro/security/CustomUserDetails.java
@@ -37,7 +37,7 @@ public class CustomUserDetails implements UserDetails {
         return user.getEmail();
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return user.getUserId();
     }
